refactor: type tier data normalizer
