+++ wikipedia/data/generate-output-files.ts
@@ -4,8 +4,14 @@
 import { execSync } from 'node:child_process';
 import * as fs from 'node:fs';
 import * as path from 'node:path';
-import { WIKIPEDIA_DATA_SOURCES } from '../config.js';
 import { normaliseLeagueTableEntry } from './output-entry-normalizer.ts';
+import {
+  isTierData,
+  normaliseOutcomeList,
+  normaliseTierData,
+  normaliseTierMetadata,
+  sanitizeRows,
+} from './output-tier-normalizer.ts';
 export { normaliseLeagueTableEntry } from './output-entry-normalizer.ts';
 
 /** @typedef {import('../models/output-file.ts').LeagueTableEntry} LeagueTableEntry */
@@ -705,120 +711,6 @@ export function buildDatasetMetadata({
 }
 
 /**
- * Ensure we have a string array with no duplicates.
- * @param {unknown} value
- * @param {LeagueTableEntry[]} fallbackRows
- * @param {'wasRelegated' | 'wasPromoted'} flag
- */
-function normaliseOutcomeList(value, fallbackRows, flag) {
-  /** @type {Set<string>} */
-  const results = new Set();
-
-  if (Array.isArray(value)) {
-    for (const entry of value) {
-      const name = toStringValue(entry);
-      if (name) results.add(name);
-    }
-  }
-
-  if (!results.size && fallbackRows.length) {
-    for (const row of fallbackRows) {
-      if (row[flag] && toStringValue(row.team)) {
-        results.add(row.team);
-      }
-    }
-  }
-
-  return Array.from(results);
-}
-
-/**
- * Remove malformed rows and ensure every row has a team name before normalisation.
- * @param {Array<Partial<LeagueTableEntry> & Record<string, unknown>>} rows
- */
-function sanitizeRows(rows) {
-  const list = Array.isArray(rows) ? rows : [];
-  /** @type {Array<Partial<LeagueTableEntry> & Record<string, unknown>>} */
-  const sanitized = [];
-
-  for (const row of list) {
-    if (!row || typeof row !== 'object') continue;
-    const teamName = toStringValue(row.team);
-    if (!teamName) continue;
-    sanitized.push({ ...row, team: teamName });
-  }
-
-  return sanitized;
-}
-
-/**
- * @param {unknown} tierValue
- * @returns {tierValue is TierData}
- */
-function isTierData(tierValue) {
-  return (
-    tierValue != null &&
-    typeof tierValue === 'object' &&
-    'season' in tierValue &&
-    'table' in tierValue
-  );
-}
-
-/**
- * @param {Record<string, unknown>} value
- */
-function normaliseTierMetadata(value) {
-  const directMetadata =
-    value.metadata && typeof value.metadata === 'object'
-      ? { .../** @type {Record<string, unknown>} */ (value.metadata) }
-      : {};
-  const legacySeasonMetadata =
-    value.seasonMetadata && typeof value.seasonMetadata === 'object'
-      ? { .../** @type {Record<string, unknown>} */ (value.seasonMetadata) }
-      : {};
-
-  const metadata = {
-    ...directMetadata,
-  };
-
-  if (value.sourceUrl != null && metadata.sourceUrl == null) {
-    metadata.sourceUrl = value.sourceUrl;
-  }
-  if (value.seasonSlug != null && metadata.seasonSlug == null) {
-    metadata.seasonSlug = value.seasonSlug;
-  }
-  if (value.tier != null && metadata.tierKey == null) {
-    metadata.tierKey = value.tier;
-  }
-  if (value.title != null && metadata.title == null) {
-    metadata.title = value.title;
-  }
-  if (legacySeasonMetadata.leagueId != null && metadata.leagueId == null) {
-    metadata.leagueId = legacySeasonMetadata.leagueId;
-  }
-  if (legacySeasonMetadata.tableIndex != null && metadata.tableIndex == null) {
-    metadata.tableIndex = legacySeasonMetadata.tableIndex;
-  }
-  if (legacySeasonMetadata.tableCount != null && metadata.tableCount == null) {
-    metadata.tableCount = legacySeasonMetadata.tableCount;
-  }
-  if (legacySeasonMetadata.seasonSlug != null && metadata.seasonSlug == null) {
-    metadata.seasonSlug = legacySeasonMetadata.seasonSlug;
-  }
-
-  if (metadata.source == null) {
-    if (metadata.leagueId != null || metadata.title != null || value.seasonMetadata != null) {
-      metadata.source = WIKIPEDIA_DATA_SOURCES.overview.sourceId;
-    } else if (metadata.sourceUrl != null || metadata.tierKey != null) {
-      metadata.source = WIKIPEDIA_DATA_SOURCES.promotion.sourceId;
-    }
-  }
-
-  const cleaned = Object.fromEntries(Object.entries(metadata).filter(([, entry]) => entry != null));
-  return Object.keys(cleaned).length ? cleaned : undefined;
-}
-
-/**
  * @param {Record<string, unknown>} seasonInfoValue
  * @param {string} seasonKey
  * @returns {SeasonInfo}
@@ -876,46 +768,6 @@ function normaliseSeasonInfo(seasonInfoValue, seasonKey) {
         : null,
     specialCompetitions,
     notes: toStringValue(seasonInfoValue.notes),
-  });
-}
-
-/**
- * @param {Record<string, unknown>} tierValue
- * @param {string} seasonKey
- */
-function normaliseTierData(tierValue, seasonKey) {
-  const table = sanitizeRows(tierValue.table);
-  const normalisedTable = table.map((row) => normaliseLeagueTableEntry(row));
-
-  const parsedSeason = Number.parseInt(String(tierValue.season ?? seasonKey), 10);
-  const fallbackSeason = Number.parseInt(seasonKey, 10);
-  const season = Number.isFinite(parsedSeason)
-    ? parsedSeason
-    : Number.isFinite(fallbackSeason)
-    ? fallbackSeason
-    : 0;
-
-  const extra = { ...tierValue };
-  delete extra.table;
-  delete extra.season;
-  delete extra.relegated;
-  delete extra.promoted;
-  delete extra.metadata;
-  delete extra.sourceUrl;
-  delete extra.seasonSlug;
-  delete extra.tier;
-  delete extra.title;
-  delete extra.seasonMetadata;
-
-  const metadata = normaliseTierMetadata(tierValue);
-
-  return /** @type {TierData} */ ({
-    ...extra,
-    season,
-    table: normalisedTable,
-    relegated: normaliseOutcomeList(tierValue.relegated, normalisedTable, 'wasRelegated'),
-    promoted: normaliseOutcomeList(tierValue.promoted, normalisedTable, 'wasPromoted'),
-    ...(metadata ? { metadata } : {}),
   });
 }
